Move ClaimTermsAndCreateAccountButton to its own file.
