+++ tests/e2e/specs/setup-mc/step-4-complete-campaign.test.js
@@ -81,11 +81,6 @@ test.describe( 'Complete your campaign', () => {
 				[ 'GET' ]
 			),
 
-			// Mock syncable products count
-			completeCampaign.fulfillSyncableProductsCountRequest( {
-				count: 1024,
-			} ),
-
 			// The following mocks are requests will happen after completing the onboarding
 			completeCampaign.mockSuccessfulSettingsSyncRequest(),
